+++ package.json
@@ -13,7 +13,8 @@
   "dependencies": {
     "@google/generative-ai": "^0.24.1",
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "react-router-dom": "^7.9.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.35.0",
