+++ src/context/AuthContext.jsx
@@ -1,0 +1,25 @@
+import React, { createContext, useState, useContext } from 'react';
+
+const AuthContext = createContext(null);
+
+export const AuthProvider = ({ children }) => {
+  const [user, setUser] = useState(JSON.parse(sessionStorage.getItem('user')));
+
+  const login = (userData) => {
+    sessionStorage.setItem('user', JSON.stringify(userData));
+    setUser(userData);
+  };
+
+  const logout = () => {
+    sessionStorage.clear();
+    setUser(null);
+  };
+
+  const value = { user, login, logout };
+
+  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
+};
+
+export const useAuth = () => {
+  return useContext(AuthContext);
+};
