+++ src/pages/login.css
@@ -1,0 +1,121 @@
+.login-container {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
+  background-color: #1a1a2e;
+}
+
+#parallax-grid {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  background-size: 60px 60px;
+  background-image:
+    linear-gradient(to right, rgba(255, 255, 255, 0.05) 1px, transparent 1px),
+    linear-gradient(to bottom, rgba(255, 255, 255, 0.05) 1px, transparent 1px);
+}
+
+#main-login {
+  position: relative;
+  padding: 2.5rem;
+  background-color: #f7f7f7;
+  border-radius: 8px;
+  box-shadow: 8px 8px 0px #e53e3e;
+  transition: all 0.2s ease-out;
+}
+
+#main-login:hover {
+  transform: translate(-2px, -2px);
+  box-shadow: 10px 10px 0px #e53e3e;
+}
+
+#login {
+  width: 32ch;
+}
+
+#login > * {
+  display: block;
+  margin-block: 1.2rem;
+}
+
+h1 {
+  color: #1a202c;
+  font-family: "Jost", sans-serif;
+  font-size: 2.2rem;
+  font-weight: 700;
+  text-align: center;
+  margin-bottom: 2rem;
+}
+
+input {
+  width: 100%;
+  border: none;
+  outline: none;
+  padding: 0.5rem 0;
+  border-bottom: 2px solid #ddd;
+  background-color: transparent;
+  color: #1a202c;
+  font-family: "roboto mono", monospace;
+  font-size: 1.1rem;
+  font-weight: 500;
+  transition: border-color 0.3s ease;
+}
+
+input:focus {
+  border-bottom-color: #e53e3e;
+}
+
+#submit {
+  border: none;
+  outline: none;
+  width: 100%;
+  background-color: #e53e3e;
+  color: #FFFFFF;
+  font-family: "Jost", sans-serif;
+  font-size: 1.1rem;
+  padding: 0.8rem;
+  margin-top: 2rem;
+  border-radius: 6px;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.2s ease-out;
+}
+
+#submit:hover {
+  transform: translateY(-3px);
+  box-shadow: 0 4px 10px rgba(229, 62, 62, 0.3);
+  background-color: #c53030;
+}
+
+#submit:active {
+  transform: translateY(-1px);
+}
+
+.link-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 1.5rem;
+  gap: 0.75rem;
+}
+
+.CP {
+  font-family: "roboto mono", monospace;
+  font-size: 0.9rem;
+  cursor: pointer;
+  text-decoration: underline;
+  color: #718096;
+  transition: color 0.2s ease-in-out;
+}
+
+.CP:hover {
+  color: #1a202c;
+}
